backend/pyproject.toml
[tool.poetry]
name = "therapybot"
version = "0.1.0"
description = "AI-powered therapy bot with Django backend and LlamaIndex RAG"
authors = ["Rasheed Martin <rasheedjmartin@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <4.0"

# Django
django = "^4.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"

# LlamaIndex and AI
llama-index = "^0.10.0"
llama-index-core = "^0.10.0"
llama-index-llms-openai = "^0.1.0"
llama-index-embeddings-huggingface = "^0.2.0"
llama-index-readers-web = "^0.1.0"

# OpenAI and ML
openai = "^1.12.0"
sentence-transformers = "^2.5.0"
torch = "^2.2.0"
transformers = "^4.38.0"

# Utilities
requests = "^2.31.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.8.0"
black = "^24.9.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "therapybot.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]